# software-engineering-2-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: "myWorld Travel API"
  description: |
    ## 🔗 **HATEOAS Architecture (REST Level 3)**
    This API implements **Hypermedia As The Engine Of Application State (HATEOAS)** following REST Level 3 architectural constraints:
    - **Dynamic Navigation**: Every response includes discoverable links to related resources and available actions
    - **Self-Descriptive**: Clients can navigate the API entirely through hypermedia links without hardcoding URLs
    - **Stateless Interaction**: All possible actions are dynamically provided based on current resource state

    Each resource response includes a `links` object containing:
    - Available HTTP methods and URIs for resource operations
    - Navigation paths to connected entities  
    - Self-reference links to the current resource

    **Important**: Link availability is dynamic and depends on the current resource state, user permissions, and context. The examples show typical links but actual responses may include additional or fewer links based on the specific situation.

    ---
    🌍 **myWorld Travel API** - Your intelligent travel companion for discovering the world!

    A comprehensive RESTful API that powers personalized travel experiences through smart recommendations and community-driven insights.

    ### 🚀 **Key Features**
    - **AI-Powered Recommendations**: Intelligent place suggestions based on your preferences, location, and travel history
...
    
    ### 👤 **User Management**
...

...

...

...

...

...
servers:
...
paths:
...
  "/users/{userId}/preference-profiles":
    get:
      summary: "Get all preference profiles for a user"
      operationId: getPreferenceProfiles
      description: |
        **FR2 - The user must be able to view their preference profiles.**

        Retrieves all preference profiles created by the user, including their names and selected categories.
      tags: [Preferences]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user whose preference profiles are being retrieved"
      responses:
        '200':
          description: "Successfully retrieved list of all preference profiles for the user"
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/PreferenceProfile'
                  links:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/HypermediaLink'
                    description: "Available actions for preference profiles collection"
                    example:
                      self:
                        href: "/users/16180/preference-profiles"
                        method: GET
                      create-profile:
                        href: "/users/16180/preference-profiles"
                        method: POST
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '404':
          description: "User not found - No user exists with the specified userId"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: USER_NOT_FOUND
                message: "User with ID 12345 not found"
    post:
      summary: "Create a new preference profile for a user"
      operationId: createPreferenceProfile
      description: |
        **FR3 - The user must be able to create preference profiles.**

        **FR4 - The user must be able to set their preferences for a profile.**

        Creates a new preference profile with selected categories for personalized recommendations. Returns all user profiles including the newly created one.
      tags: [Preferences]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user for whom the preference profile is being created"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categories]
              properties:
                name:
                  type: string
                  description: "Custom name for the preference profile to help identify its purpose"
                  example: "Cultural Explorer"
                categories:
                  type: array
                  items:
                    $ref: "#/components/schemas/PreferenceCategory"
                  uniqueItems: true
                  description: "Array of preference categories that define this profile's interests"
                  example: [MUSEUM, RESTAURANT, CULTURE]
                selectedPreferences:
                  type: array
                  items:
                    $ref: "#/components/schemas/PreferenceCategory"
                  uniqueItems: true
                  description: "Legacy alias for categories field"
                  example: [MUSEUM, RESTAURANT, CULTURE]
                  deprecated: true
                profileName:
                  type: string
                  description: "Legacy alias for name field"
                  deprecated: true
                  example: "Cultural Explorer"
      responses:
        '201':
          description: "Preference profile successfully created with the specified settings. Returns all user profiles including the newly created one."
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: "#/components/schemas/PreferenceProfile"
                    description: "All preference profiles for this user including the newly created one"
                  links:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/HypermediaLink"
                    description: "Available actions for preference profiles"
                    example:
                      self:
                        href: "/users/16180/preference-profiles"
                        method: GET
                      create-profile:
                        href: "/users/16180/preference-profiles"
                        method: POST
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '400':
          description: "Bad request - Missing or invalid profile data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: INVALID_PROFILE_DATA
                message: "No preferences selected"
                details:
                  field: categories
                  value: []
                  reason: "At least one preference category must be selected"
        '404':
          description: "User not found - No user exists with the specified userId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: USER_NOT_FOUND
                message: "User with ID 12345 not found"
  "/users/{userId}/preference-profiles/{profileId}":
    put:
      summary: "Update a specific preference profile"
      operationId: updatePreferenceProfile
      description: |
        **FR5 - The user must be able to edit their preference profiles.**

        Updates an existing preference profile with new categories and/or profile name. Returns all user profiles.
      tags: [Preferences]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user who owns the preference profile"
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the preference profile to be updated"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceProfile'
      responses:
        '200':
          description: "Preference profile successfully updated with the new settings. Returns all user profiles."
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: "#/components/schemas/PreferenceProfile"
                  links:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/HypermediaLink"
                    description: "Available actions for this user's preference profiles"
                    example:
                      self:
                        href: "/users/16180/preference-profiles"
                        method: GET
                      create-profile:
                        href: "/users/16180/preference-profiles"
                        method: POST
                      recommendations:
                        href: "/users/16180/recommendations{?latitude,longitude}"
                        method: GET
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '400':
          description: "Bad request - Invalid profile data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: INVALID_PROFILE_DATA
                message: "No preferences selected"
                details:
                  field: categories
                  value: []
                  reason: "At least one preference category must be selected"
        '404':
          description: "User or profile not found - Either the user or the specified profile does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: PROFILE_NOT_FOUND
                message: "Profile with ID 42 not found for user 12345"
        # Note: No 409 documented here because current implementation allows deleting active/last profiles.
    delete:
      summary: "Delete a specific preference profile"
      operationId: deletePreferenceProfile
      description: |
        **FR5 - The user must be able to edit their preference profiles.**

        Deletes a preference profile. **Note**: The current implementation allows deleting any profile (even the active or last profile). If the active profile is deleted, the user's `activeProfile` setting may point to a non-existent profile until changed.
      tags: [Preferences]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user who owns the preference profile"
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the preference profile to be deleted"
      responses:
        '204':
          description: "Preference profile successfully deleted with no content returned"
        '404':
          description: "User or profile not found - Either the user or the specified profile does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: PROFILE_NOT_FOUND
                message: "Profile with ID 42 not found for user 12345"
        # The spec no longer lists 409, since deletion restrictions are not enforced in code.
  "/users/{userId}/preference-profiles/{profileId}/activate":
    post:
      summary: "Activate a specific preference profile"
      operationId: activatePreferenceProfile
      description: |
        **FR5 - The user must be able to edit their preference profiles.**

        Activates a specific preference profile, making it the active profile for recommendations. Returns all user profiles with the `activeProfile` field indicating which profile is now active.
      tags: [Preferences]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user who owns the preference profiles"
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the preference profile to be activated"
      responses:
        '200':
          description: "Preference profile successfully activated. Returns all profiles with activeProfile indicator."
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: "#/components/schemas/PreferenceProfile"
                  activeProfile:
                    type: integer
                    description: "ID of the newly activated profile"
                    example: 42
                  message:
                    type: string
                    description: "Confirmation message"
                    example: "Profile activated successfully"
                  links:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/HypermediaLink"
                    description: "Available actions after activation"
                    example:
                      self:
                        href: "/users/16180/preference-profiles"
                        method: GET
                      recommendations:
                        href: "/users/16180/recommendations{?latitude,longitude}"
                        method: GET
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '404':
          description: "User or profile not found - Either the user or the specified profile does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: PROFILE_NOT_FOUND
                message: "Profile with ID 42 not found for user 12345"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User authentication using JWT token with userId, email, and role claims"
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Admin authentication using JWT token. Requires role: admin claim in the JWT payload"
  schemas:
...
    PreferenceProfile:
      type: object
      required: [profileId, userId, name, categories]
      properties:
        profileId:
          type: integer
          description: "Unique identifier for the preference profile"
          example: 42
        userId:
          type: integer
          description: "ID of the user who owns this profile"
          example: 16180
        name:
          type: string
          description: "Name of the preference profile"
          example: "Cultural Explorer"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PreferenceCategory"
          description: "List of selected preference categories for this profile"
          example: ["MUSEUM", "RESTAURANT", "CULTURE"]
        priceRange:
          type: string
          description: "Preferred price range for recommendations (if applicable)"
          example: "$$"
        # note: selectedPreferences/profileName are legacy aliases, not included here as separate fields in the profile object.
    PreferenceCategory:
      type: string
      description: "Possible categories for preferences and place classification"
      enum: [MUSEUM, BEACH, PARK, RESTAURANT, NIGHTLIFE, SHOPPING, SPORTS, CULTURE]
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: "Machine-readable error code for programmatic handling"
          example: USER_NOT_FOUND
        message:
          type: string
          description: "Human-readable error message describing the issue"
          example: "User with ID 12345 not found"
        details:
          type: object
          description: "Optional additional error details and context"
          additionalProperties: true
          example:
            field: userId
            value: 12345
            timestamp: "2025-06-01T10:30:00Z"
    HypermediaLink:
      type: object
      properties:
        href:
          type: string
          description: "URI of the related resource"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          description: "HTTP method for the link"
...